services/web/requirements.txt

flask==2.1.2

python-dotenv==0.20.0

Flask-SQLAlchemy==2.5.1

psycopg2-binary==2.9.3

gunicorn==20.1.0


# Checking pip freeze:
# (ve.gunipoc) ➜  web git:(main) ✗ pip freeze
# click==8.1.3
# Flask==2.1.2
# Flask-SQLAlchemy==2.5.1
# greenlet==1.1.2
# gunicorn==20.1.0
# itsdangerous==2.1.2
# Jinja2==3.1.2
# MarkupSafe==2.1.1
# psycopg2-binary==2.9.3
# python-dotenv==0.20.0
# SQLAlchemy==1.4.37
# Werkzeug==2.1.2

